Add tab selection shortcuts
This adds CmdOrCtrl+1 to CmdOrCtrl+8 for selecting
one of the first 8 tabs and CmdOrCtrl+9 for selecting
the last tab.  It also pulls out the shortcuts code into
its own file.
